Move FEDERAL_CODE const to misc folder
